feat(documentation): add contribution page and update migration guides
- Add new contribution documentation page with GitHub, email, and Teams contact options
- Replace ExternalTeaser component with CardTeaser across all migration guides (v1-v4)
- Add feedback teaser image asset for contribution links
- Add mail and teams teaser image assets for contact options
- Refactor CardTeaser component with new CardTeaserRow layout and styling
- Update migration guide feedback sections to link to contribution page
- Add external flag to GitHub teaser links in migration guides
- Update global styles to support new teaser component layout
