+++ content/pages/documentation/contribution/index.mdx
@@ -1,0 +1,53 @@
+---
+layout: '@template/layouts/documentation'
+sectionSpacing: 'small'
+sectionWidth: 'auto'
+title: 'Contribution'
+headline: 'Contribution'
+description: 'Get involved, share feedback, and help improve the DB UX Design System.'
+align: start
+order: 9
+toc: false
+---
+
+import CardTeaserRow from "@template/components/CardTeaser/CardTeaserRow";
+
+<CardTeaserRow
+	items={[
+				{
+			title: 'GitHub Core (Overall & Design)',
+			description:
+				'Report design issues, request features, or contribute to the design foundations.',
+			url: 'https://github.com/db-ux-design-system/core',
+			image: '/assets/teasers/github.png',
+			imageAlt: 'GitHub',
+			label: 'View on GitHub', external: true,
+		},
+		{
+			title: 'GitHub Core Web (Development)',
+			description: 'Report bugs, request features, or contribute to the component library.',
+			url: 'https://github.com/db-ux-design-system/core-web',
+			image: '/assets/teasers/github.png',
+			imageAlt: 'GitHub',
+			label: 'View on GitHub', external: true,
+		},
+		{
+			title: 'Feedback via Email',
+			description: 'Reach out to the Design System team directly via email.',
+			url: 'mailto:db-ux-designsystem@deutschebahn.com',
+			image: '/assets/teasers/mail.png',
+			imageAlt: 'Email',
+			label: 'Send Email', external: true,
+		},
+		{
+			title: 'Feedback via Teams',
+			description: 'Join the DB UX Design System channel on Microsoft Teams.',
+			url: 'https://teams.microsoft.com/l/channel/19%3A7dd9929be1fe494c823b11e0e2d22341%40thread.skype/DB%20UX%20Design%20System?groupId=72c53609-2941-46dd-8be9-e4d1d2c45c77&tenantId=a1a72d9c-49e6-4f6d-9af6-5aafa1183bfd',
+			image: '/assets/teasers/teams.png',
+			imageAlt: 'Teams',
+			label: 'Open Teams Channel', external: true,
+		},
+
+    ]}
+
+/>
