+++ content/pages/documentation/releases/migration-guides/v2.mdx
@@ -13,7 +13,7 @@ toc: true
 import { DBSection } from '@db-ux/react-core-components';
 import StatusNotification from '@template/components/StatusNotification/StatusNotification';
 import YourAction from '@template/components/YourAction/YourAction';
-import ExternalTeaserRow from '@template/components/ExternalTeaser/ExternalTeaserRow';
+import CardTeaserRow from "@template/components/CardTeaser/CardTeaserRow";
 
 <DBSection spacing="small" width="auto">
 
@@ -164,7 +164,7 @@ The following fixes require special attention. An overview of all other resolved
 </DBSection>
 <DBSection spacing="small" width="auto">
 
-<ExternalTeaserRow
+<CardTeaserRow
 	items={[
 		{
 			title: 'Design + Code Issues v2.0.0',
@@ -172,15 +172,15 @@ The following fixes require special attention. An overview of all other resolved
 			url: 'https://github.com/orgs/db-ux-design-system/projects/4/views/12?filterQuery=has%3Amilestone+-label%3A%F0%9F%96%BC%EF%B8%8Ficons+milestone%3A2.0.0',
 			image: '/assets/teasers/github.png',
 			imageAlt: 'GitHub',
-			label: 'View on GitHub',
+			label: 'View on GitHub', external: true,
 		},
 		{
 			title: 'Feedback Welcome',
 			description: 'Let us know if you encounter unexpected issues or need support.',
-			url: 'https://github.com/db-ux-design-system/core/issues/new',
-			image: '/assets/teasers/github.png',
-			imageAlt: 'GitHub',
-			label: 'Create Issue on GitHub',
+			url: '/documentation/contribution',
+			image: '/assets/teasers/feedback.png',
+			imageAlt: 'Contribution',
+			label: 'Get in touch',
 		},
 	]}
 />
