+++ content/pages/documentation/resources/index.mdx
@@ -15,10 +15,10 @@ toc: true
 
 import { DBSection, DBStack } from '@db-ux/react-core-components';
 import ResourceCards from '@template/components/ResourceCards/ResourceCards.astro';
-import ExternalTeaser from '@template/components/ExternalTeaser/ExternalTeaser';
-import ExternalTeaserRow from '@template/components/ExternalTeaser/ExternalTeaserRow';
+import CardTeaser from "@template/components/CardTeaser/CardTeaser";
+import CardTeaserRow from "@template/components/CardTeaser/CardTeaserRow";
 
-<ExternalTeaserRow
+<CardTeaserRow
 	items={[
 		{
 			title: 'Figma Access for External Employees',
