+++ template/styles/global.css
@@ -8,6 +8,7 @@
 @import url('./table.css');
 @import url('../components/GuidelineExample/GuidelineExample.css');
 @import url('../components/YourAction/YourAction.css');
+@import url('../components/CardTeaser/CardTeaserRow.css');
 @import url('./db-ux-overrides.css');
 
 :root {
